+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -2,6 +2,17 @@ import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/
 import { AsyncCartService } from 'src/app/shared/services/cart.service';
 import { Cart, CartItem } from 'src/types/cart';
 
+interface SortingOption{
+  field: string, 
+  title: string
+}
+
+const SortingOptions: SortingOption[] = [
+  { field: "amount", title: "amount"},
+  { field: "product.name", title: "name"},
+  { field: "product.price", title: "price"}
+];
+
 @Component({
   selector: 'app-cart-list',
   templateUrl: './cart-list.component.html',
@@ -12,6 +23,8 @@ export class CartListComponent {
   cartItems: CartItem[] = [];
   totalPrice: number;
   totalAmount: number;
+  sortingOptions = SortingOptions;
+  activeSortingOption = SortingOptions[0];
   constructor(
     private cartService: AsyncCartService,
     private cdr: ChangeDetectorRef